create routes/api.js and move api code to there
